migliorato script di installazione ambiente

+++ _src/_sh/_gw.environment.setup.sh
@@ -63,6 +63,21 @@ apt-get install -y php-zip
 ## installazione di php-intl
 apt-get install -y php-intl
 
+## installazione di mysql
+apt-get install -y mysql-server
+
+## installazione di adminer
+apt-get install -y adminer
+a2enconf adminer
+
+## password di root
+read -p "inserisci la password per MySQL root (vuoto per saltare): " SRVPASS
+if [ -n "$SRVPASS" ]; then
+    sudo /etc/init.d/mysql stop
+    sudo mysqld --skip-grant-tables &
+    mysql -u root mysql -e "UPDATE user SET Password=PASSWORD('$SRVPASS') WHERE User='root'; FLUSH PRIVILEGES; exit;"
+fi
+
 ## installazione di certbot
 apt-get install -y python3-certbot-apache
